refactor(contest): move problem-label helper to naming module
Address review feedback: relocate the contest problem label logic from
a private helper in contest/main.py to a public, descriptively-named
`naming.get_contest_problem_label`, alongside the other problem-naming
helpers. Tests move to test_naming.py accordingly.
