+++ rbx/box/contest/main.py
@@ -8,7 +8,7 @@ import syncer
 import typer
 
 from rbx import annotations, console, utils
-from rbx.box import cd, creation, package, presets, summary
+from rbx.box import cd, creation, naming, presets, summary
 from rbx.box.contest import contest_package, contest_state, contest_utils, statements
 from rbx.box.contest.contest_package import (
     find_contest,
@@ -245,19 +245,6 @@ def remove(path_or_short_name: str):
     )
 
 
-def _problem_command_name(problem: ContestProblem) -> str:
-    """Display label for a contest problem: '<short_name>. <problem name>'.
-
-    Falls back to just the short name when the problem package cannot be
-    loaded (e.g. missing or broken problem.rbx.yml) so the command runner
-    keeps working on a partially set-up contest.
-    """
-    pkg = package.find_problem_package(problem.get_path())
-    if pkg is not None and pkg.name:
-        return f'{problem.short_name}. {pkg.name}'
-    return problem.short_name
-
-
 @app.command(
     'each',
     help='Run a command for each problem in the contest.',
@@ -274,7 +261,7 @@ def each(ctx: typer.Context) -> None:
         CommandEntry(
             argv=argv,
             placeholder_prefix=placeholder_prefix,
-            name=_problem_command_name(problem),
+            name=naming.get_contest_problem_label(problem),
             cwd=str(problem.get_path()),
         )
         for problem in contest.problems
@@ -300,7 +287,7 @@ def on(ctx: typer.Context, problems: str) -> None:
     if len(problems_of_interest) == 1:
         command = ' '.join(['rbx'] + ctx.args)
         console.console.print(
-            f'[status]Running [item]{command}[/item] for [item]{_problem_command_name(problems_of_interest[0])}[/item]...[/status]'
+            f'[status]Running [item]{command}[/item] for [item]{naming.get_contest_problem_label(problems_of_interest[0])}[/item]...[/status]'
         )
         subprocess.call(command, cwd=problems_of_interest[0].get_path(), shell=True)
         return
@@ -310,7 +297,7 @@ def on(ctx: typer.Context, problems: str) -> None:
         CommandEntry(
             argv=argv,
             placeholder_prefix=placeholder_prefix,
-            name=_problem_command_name(p),
+            name=naming.get_contest_problem_label(p),
             cwd=str(p.get_path()),
         )
         for p in problems_of_interest
